% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-core.R
\name{add_subgraph}
\alias{add_subgraph}
\title{Add edge to a graph}
\usage{
add_subgraph(d, sg)
}
\arguments{
\item{d}{dot graph}

\item{sg}{subgraph}
}
\value{
'dot' graph
}
\description{
Add edge to a graph
}
\examples{
g <- create_graph()

# First subgraph
sg1 <- create_graph('cluster_1') |>
  add_node('a', color = 'blue', shape = 'rect') |>
  add_edge('a', 'b')

# Second subgraph
sg2 <- create_graph('cluster_2') |>
  add_node('d', style = 'filled', fillcolor = 'lightblue') |>
  add_edge('c', 'd') |>
  add_edge('c', 'e') |>
  add_edge('c', 'f') |>
  add_edge('d', 'g')    

# Add two subgraphs to main graph
g <- g |> 
  add_subgraph(sg1) |> 
  add_subgraph(sg2) |>
  add_edge('a', 'd')

g
}
