% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-out.R
\name{as.character.dot}
\alias{as.character.dot}
\title{convert a dot graph to character}
\usage{
\method{as.character}{dot}(x, ...)
}
\arguments{
\item{x}{dot graph}

\item{...}{ignored}
}
\value{
character string
}
\description{
convert a dot graph to character
}
