% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-types.R
\docType{data}
\name{dot_types}
\alias{dot_types}
\title{Some standard values for various types for different attributes}
\format{
An object of class \code{list} of length 14.
}
\usage{
dot_types
}
\description{
Some standard values for various types for different attributes
}
\keyword{datasets}
